feat & refactor: extract class definition to another file.. select team photo folder
