+++ src/IcpcResolver/UserControl/TeamDto.cs
@@ -1,13 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 
 namespace IcpcResolver.UserControl
 {
     public class TeamDto
     {
-        public int TeamId;
+        public string TeamId;
+
+        /// <summary>
+        /// team photo path
+        /// </summary>
+        public string PhotoPath;
 
         /// <summary>
         /// team's organization icon
