+++ src/IcpcResolver/Window/ContestSummary.cs
@@ -1,0 +1,14 @@
+﻿namespace IcpcResolver.Window
+{
+    public class ContestSummary
+    {
+        public string ContestLength;
+        public string FreezeTime;
+        public long PenaltyTime;
+        public string ContestName;
+        public int TeamCount;
+        public int ProblemCount;
+        public int SubmissionCount;
+        public int GroupCount;
+    }
+}
