+++ src/IcpcResolver/Window/FirstSolveInfo.cs
@@ -1,0 +1,10 @@
+﻿namespace IcpcResolver.Window
+{
+    public class FirstSolveInfo
+    {
+        public string ShortName { get; set; }
+        public string ProblemId { get; set; }
+        public string TeamId { get; set; }
+        public bool Solved { get; set; }
+    }
+}
