+++ src/IcpcResolver/Window/Organization.cs
@@ -1,0 +1,8 @@
+﻿namespace IcpcResolver.Window
+{
+    public class Organization
+    {
+        public string Id;
+        public string Name;
+    }
+}
